Update dev group seed + lint fixes
